Abstract the popover part of PopoverButton to its own compound component
